update startup messages
