+++ apiml/.gitignore
@@ -35,3 +35,11 @@ out/
 
 ### VS Code ###
 .vscode/
+
+## Caching service infinispan artifacts
+# created by infinispan during build
+index
+*.lck
+ispn12.*
+*.state
+
