+++ caching-service/src/test/java/org/zowe/apiml/caching/config/AttlsConfigTest.java
@@ -30,7 +30,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.TestPropertySource;
-import org.zowe.apiml.caching.CachingService;
+import org.zowe.apiml.caching.CachingServiceApplication;
 import org.zowe.apiml.filter.AttlsHttpHandler;
 import org.zowe.apiml.util.config.SslContext;
 
@@ -45,7 +45,7 @@ import static org.mockito.Mockito.atLeast;
 import static org.mockito.Mockito.verify;
 
 @SpringBootTest(
-    classes = CachingService.class,
+    classes = CachingServiceApplication.class,
     webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT
 )
 @ActiveProfiles("AttlsConfigTestCachingService")
